backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai==0.8.3
protobuf==4.25.3
pymongo==4.7.2
motor==3.4.0
python-dotenv==1.0.1
Pillow==10.4.0
<<<<<<< HEAD
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
=======
motor
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv
>>>>>>> 7565dd540088533d5f57a8b707d39e67a9ad78ca
